+++ src/Universum/Unsafe.hs
@@ -30,6 +30,10 @@ import Data.Maybe (fromJust)
 import Universum.Base (Int)
 import Universum.Function (flip)
 
+-- Non empty list definition for @liquidhaskell@.
+{-@ type NonEmptyList a = {xs : [a] | len xs > 0} @-}
+
 -- | Similar to '!!' but with flipped arguments.
+{-@ at :: n : Nat -> {xs : NonEmptyList a | len xs > n} -> a @-}
 at :: Int -> [a] -> a
-at = flip (!!)
+at n xs = xs !! n