+++ src/Universum/String/Reexport.hs
@@ -13,15 +13,11 @@ module Universum.String.Reexport
 
          -- * ByteString
        , module Data.ByteString
-
-         -- * Buildable
-       , module Data.Text.Buildable
        ) where
 
 import Data.ByteString (ByteString)
 import Data.String (IsString (..))
 import Data.Text (Text, lines, unlines, unwords, words)
-import Data.Text.Buildable (Buildable)
 import Data.Text.Encoding (decodeUtf8', decodeUtf8With)
 import Data.Text.Encoding.Error (OnDecodeError, OnError, UnicodeException, lenientDecode,
                                  strictDecode)